refactor: move notifier to separate file, hide internal type
