+++ packages/flame/lib/game.dart
@@ -10,6 +10,7 @@ export 'src/game/camera/camera.dart';
 export 'src/game/camera/viewport.dart';
 export 'src/game/flame_game.dart';
 export 'src/game/game.dart';
+export 'src/game/game_load_progress_notifier.dart';
 export 'src/game/game_widget/game_widget.dart';
 export 'src/game/mixins/has_draggables.dart';
 export 'src/game/mixins/has_hoverables.dart';
