+++ packages/flame/lib/src/game/game.dart
@@ -3,6 +3,7 @@ import 'dart:async';
 import 'package:flame/cache.dart';
 import 'package:flame/components.dart';
 import 'package:flame/extensions.dart';
+import 'package:flame/game.dart';
 import 'package:flame/src/flame.dart';
 import 'package:flame/src/game/game_render_box.dart';
 import 'package:flame/src/game/overlay_manager.dart';
@@ -344,38 +345,5 @@ abstract class Game {
   @internal
   void refreshWidget() {
     gameStateListeners.forEach((callback) => callback());
-  }
-}
-
-typedef LoadingFutureFactory = Future<void>? Function();
-
-class GameLoadProgressNotifier<M> {
-  StreamController<M>? loadingStreamController;
-
-  LoadingFutureFactory? _externalLoaderFuture;
-
-  void reportLoadingProgress(M message) =>
-      loadingStreamController?.add(message);
-
-  Stream<dynamic> initStreamLoader(LoadingFutureFactory loaderFutureFactory) {
-    final stream = loadingStreamController?.stream;
-    if (stream != null) {
-      return stream;
-    }
-    loadingStreamController =
-        StreamController<M>(onListen: _startOnLoadWithStream);
-    _externalLoaderFuture = loaderFutureFactory;
-    return loadingStreamController!.stream;
-  }
-
-  Future<void> _startOnLoadWithStream() async {
-    assert(_externalLoaderFuture != null);
-    await _externalLoaderFuture!.call();
-    _externalLoaderFuture = null;
-  }
-
-  void dispose() {
-    loadingStreamController?.close();
-    loadingStreamController = null;
   }
 }
